requirements.txt
Flask==3.1.0
flask-cors==5.0.0
pymongo==4.8.0
dnspython==2.6.1
python-dotenv==1.0.1
PyJWT==2.9.0
Werkzeug==3.1.3